Refactor frontmatter to Object parsing
